Use clock_settime to implement settimeofday.
Unconditionally, on all ports, use clock_settime to implement
settimeofday.  This is a little different from using clock_settime to
implement stime, because the vestigial "set time zone" feature of
settimeofday complicates matters.

The only remaining uses of this feature that aren't just bugs, are
using it to inform the Linux kernel of the offset between the hardware
clock and UTC, on systems where the hardware clock doesn't run in
UTC (usually because of dual-booting with Windows).  There currently
isn't any other way to do this.  However, the callers that do this
call settimeofday with _only_ the timezone argument non-NULL.
Therefore, glibc's new behavior is: callers of settimeofday must
supply one and only one of the two arguments.  If both arguments are
non-NULL, or both arguments are NULL, the call fails and sets errno to
EINVAL.

When only the timeval argument is supplied, settimeofday calls
__clock_settime(CLOCK_REALTIME), same as stime.

When only the timezone argument is supplied, settimeofday calls a new
internal function called __settimezone.  On Linux, only, this function
will pass the timezone structure to the settimeofday system call.  On
all other operating systems, and on Linux architectures that don't
define __NR_settimeofday, __settimezone is a stub that always sets
errno to ENOSYS and returns -1.

The same semantics are implemented for Linux/Alpha's GLIBC_2.0 compat
symbol for gettimeofday.

There are no longer any internal callers of __settimeofday, so the
internal prototype is removed.

	* time/settimeofday.c (settimeofday): No longer a stub
	implementation.  Call __clock_settime or __settimezone depending
	on arguments.  Optionally override the default symbol version for
	settimeofday.
	* include/sys/time.h: Remove prototype for __settimeofday.
	Add prototype for __settimezone.
	* sysdeps/unix/syscalls.list: Remove entry for settimeofday.

	* time/settimezone.c: New file.
	(__settimezone): New stub implementation.
	* sysdeps/unix/sysv/linux/settimezone.c: New file.
	(__settimezone): Implement using settimeofday system call,
	if available.
	* time/Makefile (routines): Add settimezone.

	* sysdeps/unix/sysv/linux/alpha/syscalls.list:
	Remove entries for settimeofday and osf_settimeofday.
	* sysdeps/unix/sysv/linux/alpha/osf_settimeofday.c
	New file, defines settimeofday@GLIBC_2.0.
	* sysdeps/unix/sysv/linux/alpha/settimeofday.c:
	New file, defines settimeofday@@GLIBC_2.1.