Use clock_gettime to implement gettimeofday.
Abstractly, this is the same change as using clock_gettime to
implement time, but the vestigial "get time zone" feature of
gettimeofday complicates matters a little.  Unlike settimeofday, there
are *no* known uses of this feature that are not bugs, so it is simply
dummied out.  (The per-process timezone support in localtime and
friends is unrelated, and the programs that set the kernel's offset
between the hardware clock and UTC do not need to read it back.)

Henceforth, if gettimeofday's "struct timezone" argument is not NULL,
it will write zeroes to both fields.  Any program that is actually
looking at this data will thus think it is running in UTC, which is
probably more correct than whatever it was doing before.

Hurd having already been converted, this patch only affects Linux and
hypothetical future ports.  Most Linux ports supplied a vDSO symbol
for gettimeofday, and some wrapped that with an ifunc, so this patch
deletes a lot of code.  For ease of future edits to the many copies of
_libc_vdso_platform_setup, the variable "p" in each is now declared
separately from any use of it.

As with settimeofday, the alpha-linux-gnu configuration has two
versions, GLIBC_2.0 and GLIBC_2.1, with the older symbol using 32-bit
time_t (yes, really) and the same solution is implemented here.

__gettimeofday no longer has any internal callers, so remove its
internal prototype.  The __gettimeofday@GLIBC_2.0 export remains, in
case it is used by any third-party code.

	* time/gettimeofday.c: No longer a stub implementation.
	Call __clock_gettime.  If tz argument is not NULL, clear the
	object it points to.  Remove libc_hidden_def for __gettimeofday
	and libc_hidden_weak for gettimeofday.  Optionally override the
	default symbol version for gettimeofday.
	* include/sys/time.h: Remove internal prototype and libc_hidden_proto
        for __gettimeofday, and libc_hidden_proto for gettimeofday.

	* sysdeps/unix/sysv/linux/alpha/gettimeofday.c:
	New file, uses time/gettimeofday.c but changes the default
	symbol version to gettimeofday@@GLIBC_2.1.

	* sysdeps/unix/sysv/linux/alpha/osf_gettimeofday.c
        (gettimeofday@GLIBC_2.0): Use __clock_gettime instead of __gettimeofday.
        If tz argument is not NULL, clear the object it points to.

	* sysdeps/unix/sysv/linux/aarch64/init-first.c
	* sysdeps/unix/sysv/linux/arm/init-first.c
	* sysdeps/unix/sysv/linux/mips/init-first.c
	* sysdeps/unix/sysv/linux/powerpc/init-first.c
	* sysdeps/unix/sysv/linux/riscv/init-first.c
	* sysdeps/unix/sysv/linux/s390/init-first.c
	* sysdeps/unix/sysv/linux/sparc/init-first.c:
	Do not define nor initialize VDSO_SYMBOL(gettimeofday).

	* sysdeps/unix/sysv/linux/aarch64/libc-vdso.h
	* sysdeps/unix/sysv/linux/arm/libc-vdso.h
	* sysdeps/unix/sysv/linux/mips/libc-vdso.h
	* sysdeps/unix/sysv/linux/powerpc/libc-vdso.h
	* sysdeps/unix/sysv/linux/riscv/libc-vdso.h
	* sysdeps/unix/sysv/linux/s390/libc-vdso.h
	* sysdeps/unix/sysv/linux/sparc/libc-vdso.h:
	Do not declare VDSO_SYMBOL(gettimeofday).

	* sysdeps/unix/syscalls.list: Remove entry for gettimeofday.
	* sysdeps/unix/sysv/linux/alpha/syscalls.list:
	Remove entries for gettimeofday and osf_gettimeofday.
	* sysdeps/unix/sysv/linux/x86_64/x32/syscalls.list:
	Remove entry for gettimeofday.

	* sysdeps/posix/gettimeofday.c
	* sysdeps/unix/sysv/linux/aarch64/gettimeofday.c
	* sysdeps/unix/sysv/linux/gettimeofday.c
	* sysdeps/unix/sysv/linux/i386/gettimeofday.c
	* sysdeps/unix/sysv/linux/powerpc/gettimeofday.c
	* sysdeps/unix/sysv/linux/x86/gettimeofday.c: Delete file.

	* sysdeps/unix/make-syscalls.sh: Change an example in a comment
	from referring to gettimeofday, to referring to sigaction.