Use clock_gettime to implement timespec_get.
timespec_get is the same function as clock_gettime, with an obnoxious
coating of NIH painted on it by the ISO C committee.  In addition to
the rename, it takes its arguments in a different order, it returns 0
on *failure* or a positive number on *success*, and it places
requirements on the numeric values of its TIME_* constants that make
it impossible for us to reuse the CLOCK_* constants we already have
for it.  As such I think the most sensible GNUish policy for this
function is that we will provide it, but we will not recommend its
use, we will not go out of our way to make it efficient, and we will
not attempt to extend the set of TIME_* constants beyond the ones
defined in ISO C (and in POSIX, if it ever appears in POSIX).

This patch simply promotes the sysdeps/posix implementation to
universal, and removes the Linux-specific implementation, whose
apparent reason for existing was to cut out one function call's worth
of overhead.  Oh well.

	* sysdeps/posix/timespec_get.c
	* sysdeps/unix/sysv/linux/timespec_get.c: Delete file.
	* time/timespec_get.c: No longer a stub.  Replace implementation
	with the code formerly in sysdeps/posix/timespec_get.c.